test(runtime-scanner): move Node.js policy tests in FSRT
