+++ crates/fsrt/Cargo.toml
@@ -33,3 +33,6 @@ walkdir.workspace = true
 graphql-parser = "0.4.0"
 glob = "0.3.2"
 regex.workspace = true
+
+[dev-dependencies]
+time.workspace = true
